requirements_video.txt
moviepy>=2.0.0
Pillow>=9.0.0
numpy>=1.23.0
